refactor: Inline method proxy into resource proxy
